# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Forge local control API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Local-only HTTP/JSON and SSE control contract for Forge.

    This API is served over an owner-only Unix domain socket, not a TCP
    listener. The nominal server URL only supplies an HTTP authority to generic
    OpenAPI clients. M0 has no application-level authentication: the local
    process is the sole operator. Every mutation is a named command, committed
    by Core, and requires an `Idempotency-Key` header plus an expected Project
    revision. The operator is recorded in canonical Events as a `human` actor;
    M0 defines no separate local-administrator actor kind.
servers:
  - url: http://forge.local
    description: Nominal HTTP authority carried over the local Unix socket.
...
security: []
...
paths:
...
  /v1/projects/{project_id}/knowledge:
    get:
      tags: [knowledge]
      operationId: listKnowledgePages
      summary: List current canonical page revisions, including drafts and withdrawn pages.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: after, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Ascending page-ID order; next_cursor is the last returned page ID when more pages exist.
          content: {application/json: {schema: {$ref: '#/components/schemas/KnowledgePageList'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/knowledge/{page_id}:
    get:
      tags: [knowledge]
      operationId: getKnowledgePage
      summary: Read the current canonical page revision.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: page_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
      responses:
        '200':
          description: Current page, including draft or withdrawn state.
          content: {application/json: {schema: {$ref: '#/components/schemas/KnowledgePage'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/knowledge/{page_id}/history:
    get:
      tags: [knowledge]
      operationId: listKnowledgePageHistory
      summary: Read immutable revisions of one canonical page.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: page_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: after, in: query, schema: {type: integer, minimum: 0, maximum: 9223372036854775807, format: int64}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Ascending page revision order, strictly after the supplied revision.
          content: {application/json: {schema: {$ref: '#/components/schemas/KnowledgePageHistory'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /v1/projects/{projectId}/tasks/{taskId}/file-snapshots:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TaskId'
      - $ref: '#/components/parameters/PageCursor'
      - $ref: '#/components/parameters/PageLimit'
    get:
      tags: [tasks]
      operationId: listTaskFileSnapshots
      summary: Read capture operations and immutable manifests without host paths.
      responses:
        '200':
          description: Capture status is separate from Task lifecycle. Only sealed operations have an Artifact.
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/FileSnapshotView'}}
                  next_cursor: {type: [string, 'null']}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/projects/{projectId}/tasks/{taskId}/file-inputs:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [tasks]
      operationId: listTaskFileInputs
      summary: Read immutable snapshots attached to future Runs of this Task.
      responses:
        '200':
          description: At most eight same-Project snapshots; existing RunSpecs never change.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, maxItems: 8, items: {$ref: '#/components/schemas/TaskFileInput'}}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /v1/projects/{projectId}/employees/{employeeId}/threads:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: employeeId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UuidV7'
      - name: after
        in: query
        description: Exclusive thread UUID from next_cursor; creation order is stable.
        schema:
          $ref: '#/components/schemas/UuidV7'
      - $ref: '#/components/parameters/PageLimit'
    get:
      tags: [inbox]
      operationId: listEmployeeThreads
      summary: Read a bounded page of an Employee's conversations.
      responses:
        '200':
          description: Canonical threads, including taskless conversations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeThreadList'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/projects/{projectId}/threads/{threadId}/messages:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: threadId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UuidV7'
      - name: after
        in: query
        description: Exclusive per-thread sequence, zero by default.
        schema:
          type: integer
          minimum: 0
          maximum: 9223372036854775807
      - $ref: '#/components/parameters/PageLimit'
    get:
      tags: [inbox]
      operationId: listEmployeeMessages
      summary: Read immutable messages in thread sequence order.
      responses:
        '200':
          description: Canonical messages; reading does not acknowledge them.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeMessageList'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV7'
...
    PageCursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous list response.
      schema:
        type: string
        minLength: 1
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    InvalidRequest:
      description: The request cannot be parsed or violates a transport contract.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource does not exist in the Project scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unavailable:
      description: A required local Forge dependency is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UuidV7:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      description: A UUIDv7 identifier generated by Forge unless stated otherwise.
    Rfc3339Timestamp:
      type: string
      format: date-time
      description: UTC timestamp serialized in RFC 3339 format.
    JsonObject:
      type: object
      additionalProperties: true
      description: A JSON object interpreted only by the selected domain contract.
...
    StableKey:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[a-z][a-z0-9_]{0,63}$'
      description: >
        A durable Project-defined key: lowercase ASCII, beginning with a letter,
        then lowercase ASCII letters, digits, or underscores. Display names are
        separate fields and must not be used where this schema is required.
...
    KnowledgeSourceRef:
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind, artifact_id]
          properties:
            kind: {const: artifact}
            artifact_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, event_id]
          properties:
            kind: {const: event}
            event_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, handoff_id]
          properties:
            kind: {const: task_handoff}
            handoff_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, page_id, revision]
          properties:
            kind: {const: knowledge_page}
            page_id: {$ref: '#/components/schemas/UuidV7'}
            revision: {type: integer, minimum: 1, format: int64}
    KnowledgePageContent:
      type: object
      additionalProperties: false
      required: [title, markdown, source_refs]
      properties:
        title: {type: string, minLength: 1, x-forge-max-utf8-bytes: 200}
        markdown: {type: string, minLength: 1, x-forge-max-utf8-bytes: 65536}
        source_refs:
          type: array
          maxItems: 128
          uniqueItems: true
          items: {$ref: '#/components/schemas/KnowledgeSourceRef'}
...
    KnowledgePage:
      type: object
      additionalProperties: false
      required: [id, project_id, revision, kind, status, content, content_hash, actor, command_id, created_at, revised_at, supersedes_revision]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        revision: {type: integer, minimum: 1, format: int64}
        kind: {type: string, enum: [introduction, architecture, guide, policy, decision]}
        status: {type: string, enum: [draft, published, withdrawn]}
        content: {$ref: '#/components/schemas/KnowledgePageContent'}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        actor: {$ref: '#/components/schemas/ActorReference'}
        command_id: {$ref: '#/components/schemas/UuidV7'}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        revised_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        supersedes_revision: {type: [integer, 'null'], minimum: 1, format: int64}
      description: Only published Policy and Decision pages carry managed authority.
    KnowledgePageList:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/KnowledgePage'}}
        next_cursor: {type: [string, 'null'], format: uuid}
    KnowledgePageHistory:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/KnowledgePage'}}
        next_cursor: {type: [string, 'null'], pattern: '^[0-9]+$'}
...
    SnapshotFile:
      type: object
      additionalProperties: false
      required: [path, size_bytes, executable, sha256]
      properties:
        path: {type: string, minLength: 1, maxLength: 1024}
        size_bytes: {type: integer, minimum: 0, maximum: 16777216}
        executable: {type: boolean}
        sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
      description: Safe HTTP projection; internal object-storage keys are omitted.
    FileSnapshotManifest:
      type: object
      additionalProperties: false
      required: [schema_version, project_id, source_task_id, files]
      properties:
        schema_version: {type: integer, const: 1}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        source_task_id: {$ref: '#/components/schemas/UuidV7'}
        files: {type: array, minItems: 1, maxItems: 64, items: {$ref: '#/components/schemas/SnapshotFile'}}
    TaskFileInput:
      type: object
      additionalProperties: false
      required: [artifact_id, manifest]
      properties:
        artifact_id: {$ref: '#/components/schemas/UuidV7'}
        manifest: {$ref: '#/components/schemas/FileSnapshotManifest'}
    FileSnapshotView:
      type: object
      additionalProperties: false
      required: [id, artifact_id, task_id, title, state, error_code, manifest, created_at]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        artifact_id: {$ref: '#/components/schemas/UuidV7'}
        task_id: {$ref: '#/components/schemas/UuidV7'}
        title: {type: string}
        state: {type: string, enum: [pending, sealed, failed]}
        error_code: {type: [string, 'null']}
        manifest:
          oneOf:
            - $ref: '#/components/schemas/FileSnapshotManifest'
            - {type: 'null'}
        created_at: {type: string, format: date-time}
...
    TaskMessageContext:
      type: object
      additionalProperties: false
      required: [task_id, pipeline_version_id, stage_id, stage_visit]
      properties:
        task_id:
          $ref: '#/components/schemas/UuidV7'
        pipeline_version_id:
          $ref: '#/components/schemas/UuidV7'
        stage_id:
          $ref: '#/components/schemas/StableKey'
        stage_visit:
          type: integer
          minimum: 1
    EmployeeThreadList:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            type: object
            required: [id, project_id, employee_id, task_id, revision, last_sequence, created_by, created_at, updated_at]
            properties:
              id:
                $ref: '#/components/schemas/UuidV7'
              project_id:
                $ref: '#/components/schemas/UuidV7'
              employee_id:
                $ref: '#/components/schemas/UuidV7'
              task_id:
                oneOf:
                  - $ref: '#/components/schemas/UuidV7'
                  - type: 'null'
              revision:
                type: integer
                minimum: 1
              last_sequence:
                type: integer
                minimum: 0
        next_cursor:
          type: [string, 'null']
          description: Exclusive UUID passed as after to the next request.
    EmployeeMessageList:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            type: object
            required: [id, project_id, employee_id, thread_id, sequence, sender, target, kind, requirement, body, reply_to, created_at]
            properties:
              id:
                $ref: '#/components/schemas/UuidV7'
              thread_id:
                $ref: '#/components/schemas/UuidV7'
              sequence:
                type: integer
                minimum: 1
              target:
                $ref: '#/components/schemas/MessageTarget'
              body:
                type: string
              kind:
                type: string
                enum: [instruction, question, notification, reply]
              requirement:
                type: string
                enum: [informational, acknowledged, answered]
        next_cursor:
          type: [string, 'null']
          description: Exclusive sequence passed as after to the next request.
...
    MessageTarget:
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind]
          properties:
            kind:
              const: inbox
        - type: object
          additionalProperties: false
          required: [kind, context]
          properties:
            kind:
              const: task_execution
            context:
              $ref: '#/components/schemas/TaskMessageContext'
        - type: object
          additionalProperties: false
          required: [kind, context, run_id, fencing_token, environment_epoch]
          properties:
            kind:
              const: exact_run
            context:
              $ref: '#/components/schemas/TaskMessageContext'
            run_id:
              $ref: '#/components/schemas/UuidV7'
            fencing_token:
              type: integer
              minimum: 1
            environment_epoch:
              type: integer
              minimum: 1
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - not_found
            - conflict
            - stale_revision
            - idempotency_conflict
            - forbidden
            - execution_stopped
            - cursor_invalid
            - unavailable
            - internal
        message:
          type: string
        details:
          $ref: '#/components/schemas/JsonObject'
        request_id:
          $ref: '#/components/schemas/UuidV7'
...
    ActorReference:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind:
          type: string
          enum: [human, employee, system_manager, core, supervisor]
        id:
          $ref: '#/components/schemas/UuidV7'